# lambda/openapi.json
{"openapi": "3.0.3", "info": {"title": "Powertools API", "version": "1.0.0"}, "servers": [{"url": "/"}], "paths": {"/list_of_items": {"post": {"summary": "POST /list_of_items", "description": "receives a json array made up of json objects, maps each object to a pydantic model called Product and returns the json array", "operationId": "list_of_items_list_of_items_post", "parameters": [{"description": "The json array made up of json objects. This represents an item in this ecommerce application", "required": true, "schema": {"items": {}, "type": "array", "title": "List Items", "description": "The json array made up of json objects. This represents an item in this ecommerce application", "example": {"PK": "PRODUCT", "SK": "PRODUCT#4c1fadaa-213a-4ea8-aa32-58c217604e3c", "productId": "4c1fadaa-213a-4ea8-aa32-58c217604e3c", "category": "fruit", "createdDate": "2017-04-17T01:14:03 -02:00", "description": "Culpa non veniam deserunt dolor irure elit cupidatat culpa consequat nulla irure aliqua.", "modifiedDate": "2019-03-13T12:18:27 -01:00", "name": "Fresh Lemons", "package": {"height": 948, "length": 455, "weight": 54, "width": 905}, "pictures": ["https://img.freepik.com/free-photo/lemon_1205-1667.jpg?w=1480&t=st=1689112951~exp=1689113551~hmac=196483001817bd24a3d1eeb35a23ddf9911ac5628fe6df0758a47faa7ed3e332"], "price": 7160, "tags": ["mollit", "ad", "eiusmod", "irure", "tempor"]}}, "name": "list_items", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {}, "type": "array", "title": "Return", "description": "returns list of items"}}}}}}}, "/populate_db": {"post": {"summary": "POST /populate_db", "description": "Populates the database with a list of products gotten from a json file", "operationId": "add_products_db_populate_db_post", "parameters": [{"description": "The json array made up of json objects. This represents an item in this ecommerce application", "required": true, "schema": {"items": {}, "type": "array", "title": "List Items", "description": "The json array made up of json objects. This represents an item in this ecommerce application", "example": {"PK": "PRODUCT", "SK": "PRODUCT#4c1fadaa-213a-4ea8-aa32-58c217604e3c", "productId": "4c1fadaa-213a-4ea8-aa32-58c217604e3c", "category": "fruit", "createdDate": "2017-04-17T01:14:03 -02:00", "description": "Culpa non veniam deserunt dolor irure elit cupidatat culpa consequat nulla irure aliqua.", "modifiedDate": "2019-03-13T12:18:27 -01:00", "name": "Fresh Lemons", "package": {"height": 948, "length": 455, "weight": 54, "width": 905}, "pictures": ["https://img.freepik.com/free-photo/lemon_1205-1667.jpg?w=1480&t=st=1689112951~exp=1689113551~hmac=196483001817bd24a3d1eeb35a23ddf9911ac5628fe6df0758a47faa7ed3e332"], "price": 7160, "tags": ["mollit", "ad", "eiusmod", "irure", "tempor"]}}, "name": "list_items", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "boolean", "title": "Return", "description": "Whether the products were added to the successfully"}}}}}}}, "/payment_link": {"get": {"summary": "GET /payment_link", "description": "Creates a stripe payment link", "operationId": "payment_link_payment_link_get", "parameters": [{"description": "The Product name", "required": true, "schema": {"type": "string", "title": "Product Name", "description": "The Product name"}, "name": "product_name", "in": "query"}, {"description": "The Product quantity", "required": true, "schema": {"type": "integer", "title": "Qty", "description": "The Product quantity"}, "name": "qty", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Return"}}}}}}}, "/current_time": {"get": {"summary": "GET /current_time", "description": "Gets the current time in seconds", "operationId": "current_time_current_time_get", "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "integer", "title": "Return"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}